Added cast to remove warning about %lx pointer printing.
git-svn-id: file:///Users/kling/src/protoman/svn/trunk/ytalk@80 8120b729-f3fe-0310-9465-b9035d29392c

+++ src/mem.c
@@ -167,7 +167,7 @@ clear_all()
 #endif
     while(glist != NULL) {
 #ifdef YTALK_DEBUG
-	printf("0x%lx: %d\t(%s:%d)\n", glist->addr, glist->size,
+	printf("0x%lx: %d\t(%s:%d)\n", (long unsigned)glist->addr, glist->size,
 				       glist->file, glist->line
 	);
 #endif
